+++ changelog.md
@@ -9,8 +9,10 @@
 - Added **Badge API** back as a dependency
 - Added **Most Sends** for Admins and Mods in the Sent Layouts section.
 - Added **Classic & Platformer** filters on the Sent Layouts section.
+- Added More **Nameplates**
 - Fixed issue where rubies isn't rewarded correctly at the End Level Layer.
 - Fixed issue where the **Disable Reward Animation** option isn't applying on rubies collection.
+- Fixed issue with the Menu Buttons in the **Profile Page** got cut off.
 
 # v0.4.1-beta.1